requirements.txt
ipython==8.14.0
pandas==1.4.4
python-dotenv==1.0.0
tabulate==0.9.0
